requirements.txt
opencv-python-headless==4.10.0.84
Pillow>=10,<12
numpy>=1.26,<3
